refactor(#578): dissolve workflows/email/lib → co-locate as *-lib.ts next to send-*.ts
Final slice of #578. Move the 4 email-template-data helpers out of the
workflows/email/lib/ subdir to siblings of their send-*.ts workflow, renamed
with the *-lib.ts suffix (the agreed convention for helper files that aren't a
plain module-root move):

  lib/order-canceled-customer-email.ts  -> workflows/order-canceled-customer-email-lib.ts
  lib/partner-production-run-email.ts   -> workflows/partner-production-run-email-lib.ts
  lib/partner-task-email.ts             -> workflows/partner-task-email-lib.ts
  lib/region-request-admin-email.ts     -> workflows/region-request-admin-email-lib.ts

lib/ and lib/workflows/email/ are the same depth under email/, so the helpers
have no internal relative imports to adjust. Importers updated: 3 send-*.ts
workflows (../lib/X -> ./X-lib), 2 subscribers, 1 store API route, 2 integration
specs, and the moved unit specs. Stale comment refs in the analytics digest
helpers updated. Pure import-path refactor, no behaviour change.

Verify: 37 email unit tests pass; tsc baseline unchanged (69, 0 new).
This completes #578 (module lib/ dirs landed in PR #650).
